requirements.txt
streamlit>=1.28.0
groq==0.34.1
numpy>=2.0.0
pypdf==6.3.0
python-dotenv==1.2.1
qdrant_client==1.16.0
sentence_transformers==5.1.2
python-docx>=1.1.0
python-pptx>=0.6.23
requests>=2.31.0
requests-ntlm>=1.2.0
msal>=1.24.0